feat: adapt iam service and vehicle service

+++ src/main/java/com/pe/platform/interaction/domain/services/FavoriteCommandService.java
@@ -1,10 +1,10 @@
 package com.pe.platform.interaction.domain.services;
 
-import com.pe.platform.interaction.domain.model.aggregates.Favorite;
-
 import java.util.Optional;
 
+import com.pe.platform.interaction.domain.model.aggregates.Favorite;
+
 public interface FavoriteCommandService {
-    Optional<Favorite> addFavorite(int vehicleId, long profileId);
-    void removeFavorite(int vehicleId, long profileId);
+    Optional<Favorite> addFavorite(Long vehicleId, long profileId);
+    void removeFavorite(Long vehicleId, long profileId);
 }
